requirements.txt
mysql-connector-python
sqlalchemy
streamlit
